pyproject.toml
[project]
name = "agentcore-agents"
version = "0.1.0"
description = "Agentcore agents"
readme = "README.md"
authors = [
    {name = "Benito Martin"}
]
license = {text = ""}
requires-python = ">=3.13"

dependencies = [
    "bedrock-agentcore>=1.1.1",
    "bedrock-agentcore-starter-toolkit>=0.2.2",
    "boto3>=1.42.4",
    "loguru>=0.7.3",
    "nest-asyncio>=1.6.0",
    "playwright>=1.56.0",
    "pydantic>=2.12.5",
    "strands-agents>=1.19.0",
    "strands-agents-tools>=0.2.17",
]

[dependency-groups]
dev = [
    "prek>=0.2.12",
]
lint = [
    "mypy>=1.18.2",
    "ruff>=0.14.2",
]
test = [
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentcore_agents"]


######################################
# --- Commitizien --- #
######################################
[tool.commitizen]
name = "cz_conventional_commits"   # enforce Conventional Commits style
tag_format = "v$version"           # git tags look like v1.2.3
version_scheme = "pep440"          # use PEP 440 versions (1.2.3, 1.2.3rc1, etc.)
version_provider = "uv"            # delegate version reading/bumping to uv
changelog_start_rev = "v1.0.0"     # changelog will include commits starting after v1.0.0
major_version_zero = true          # for versions < 1.0.0, feat -> minor, fix -> patch
update_changelog_on_bump = true    # auto-update CHANGELOG.md on each version bump

######################################
# --- Bandit --- #
######################################
[tool.bandit]
targets = ["src"]

######################################
# --- Linting & Formatting Tools --- #
######################################

[tool.ruff]
# Assume Python 3.13
target-version = "py313"

# Same as Black.
line-length = 100     # Line length of 100 chars (like Black default is 88)
indent-width = 4      # Indent with 4 spaces (like Black)

# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
]

# Whether to show an enumeration of all fixed lint violations
show-fixes = true

# Rules to enable
lint.select = [
  "B",      # flake8-bugbear: finds likely bugs and design problems
  "E",      # pycodestyle: PEP8 errors
  "F",      # Pyflakes: logical errors
  "I",      # isort: import order issues
  "SIM",    # similarity: detects duplicate code
  "UP",     # pyupgrade: detects outdated syntax
#   "D102",   # Missing docstring for method
#   "D103",   # Missing docstring for function
  "D414",   # Docstring missing section
  "D419",   # Empty docstring
  # "D101", # Uncomment to enforce class docstrings
]


# Rules to ignore (none by default)
lint.ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
lint.fixable = ["ALL"]

# Which rules cannot be fixed automatically
lint.unfixable = []

[tool.ruff.lint.mccabe]
max-complexity = 10   # Maximum McCabe complexity allowed per function


#########################
# --- Static Typing --- #
#########################

[tool.mypy]
files = ["src"]                  # Check all files in src/
explicit_package_bases = true    # Use explicit package bases to avoid module name conflicts

ignore_missing_imports = true    # Ignore third-party imports without type hints
disallow_untyped_defs = true     # Require type annotations for all function definitions
check_untyped_defs = true        # Check bodies of functions even if no type annotations
warn_redundant_casts = true      # Warn about redundant casts
warn_unused_ignores = false      # Warn about unused # type: ignore comments
warn_return_any = false          # Warn if a function returns Any type
strict_optional = true           # Enable strict optional checking
